Crea módulo para las funciones de localStorage
