+++ src/index.js
@@ -1,5 +1,6 @@
 import { validarPokemonABuscar } from "./validaciones/validaciones.js";
 import { buscarPokemonEspecifico, hacerSolicitud } from "./api/pokedex.js";
+import { guardarListadoPokemonesEnLocalStorage, guardarDataPokemonEnLocalStorage, cargarListadoPokemonesDeLocalStorage, cargarDataPokemonDeLocalStorage } from "./storage/pokedex.js";
 
 // Variables globales
 
@@ -83,7 +84,7 @@ function imprimirStatsPokemonElegido(infoPokemon) {
 
 function imprimirHabilidadesPokemon(infoPokemon) {
   const $contenedorHabilidadesPokemon = document.querySelector(".lista-habilidades");
-  
+
   while ($contenedorHabilidadesPokemon.firstChild) {
     $contenedorHabilidadesPokemon.removeChild($contenedorHabilidadesPokemon.firstChild);
   }
@@ -216,35 +217,6 @@ async function iniciarPagina(INDICADOR_PAGINA = 0) {
   gestionarListadoPokemones(INDICADOR_PAGINA);
 }
 iniciarPagina();
-
-async function gestionarListadoPokemones(indicadorPagina) {
-  const paginaActual = Number(document.querySelector(".active").textContent);
-  const LIMITE_POKEMONES = 9;
-
-  try {
-    const dataPokemon = cargarListadoPokemonesDeLocalStorage(paginaActual);
-    let nombresPokemon = [];
-
-    dataPokemon["results"].forEach(pokemon => {
-      const nombrePokemon = pokemon.name;
-      nombresPokemon.push(nombrePokemon);
-    });
-
-    imprimirNombresPokemon(nombresPokemon);
-  } catch (error) {
-    const dataPokemon = await hacerSolicitud(indicadorPagina, LIMITE_POKEMONES);
-    guardarListadoPokemonesEnLocalStorage(paginaActual, dataPokemon);
-
-    let nombresPokemon = [];
-
-    dataPokemon["results"].forEach(pokemon => {
-      const nombrePokemon = pokemon.name;
-      nombresPokemon.push(nombrePokemon);
-    });
-
-    imprimirNombresPokemon(nombresPokemon);
-  }
-}
 
 // DOM
 
@@ -403,6 +375,35 @@ function calcularNumeroPokemonListado(indicadorPaginaASolicitar) {
   return resultado;
 }
 
+async function gestionarListadoPokemones(indicadorPagina) {
+  const paginaActual = Number(document.querySelector(".active").textContent);
+  const LIMITE_POKEMONES = 9;
+
+  try {
+    const dataPokemon = cargarListadoPokemonesDeLocalStorage(paginaActual);
+    let nombresPokemon = [];
+
+    dataPokemon["results"].forEach(pokemon => {
+      const nombrePokemon = pokemon.name;
+      nombresPokemon.push(nombrePokemon);
+    });
+
+    imprimirNombresPokemon(nombresPokemon);
+  } catch (error) {
+    const dataPokemon = await hacerSolicitud(indicadorPagina, LIMITE_POKEMONES);
+    guardarListadoPokemonesEnLocalStorage(paginaActual, dataPokemon);
+
+    let nombresPokemon = [];
+
+    dataPokemon["results"].forEach(pokemon => {
+      const nombrePokemon = pokemon.name;
+      nombresPokemon.push(nombrePokemon);
+    });
+
+    imprimirNombresPokemon(nombresPokemon);
+  }
+}
+
 /* Buscador */
 
 const $botonBuscarPokemon = document.querySelector(".boton-buscar-pokemon");
@@ -419,49 +420,3 @@ $botonBuscarPokemon.addEventListener("click", () => {
     gestionarBusquedaPokemonEspecifico(pokemonABuscar);
   }
 })
-
-/* localStorage */
-
-function guardarListadoPokemonesEnLocalStorage(indicadorPagina, dataPokemones) {
-  if (typeof indicadorPagina !== "number" || typeof dataPokemones !== "object") {
-    throw new Error("Se necesita el número de la página a la que pertenece la data, y los pokemones para guardar en el localStorage");
-  }
-
-  localStorage.setItem(`pagina_${indicadorPagina}`, JSON.stringify(dataPokemones));
-}
-
-function cargarListadoPokemonesDeLocalStorage(indicadorPagina) {
-  if (indicadorPagina === undefined) {
-    throw new Error("Se necesita una cantidad y un indicador de página para cargar a los pokemones");
-  }
-
-  const pokemones = JSON.parse(localStorage.getItem(`pagina_${indicadorPagina}`));
-
-  if (pokemones === null) {
-    throw new Error(`Pagina ${indicadorPagina} de Pokemones no se encontró en el localStorage`);
-  }
-
-  return pokemones;
-}
-
-function guardarDataPokemonEnLocalStorage(dataPokemon) {
-  if (typeof dataPokemon !== "object") {
-    throw new Error("Se necesita la data del pokémon para guardarla en el localStorage");
-  }
-
-  localStorage.setItem(`pokemon_${dataPokemon.nombre}`, JSON.stringify(dataPokemon));
-}
-
-function cargarDataPokemonDeLocalStorage(nombrePokemon){
-	if(nombrePokemon === undefined){
-		throw new Error("Se necesita el nombre para cargar el pokémon correspondiente");
-	}
-
-	const dataPokemon = JSON.parse(localStorage.getItem(`pokemon_${nombrePokemon}`));
-
-	if(dataPokemon === null){
-		throw new Error(`Pokémon ${nombrePokemon} no se encontró en el localStorage`);
-	}
-
-	return dataPokemon;
-}
